Bumped load_time_intervals from SectorReader into SosModelReader
